Scripts/Elwynn moved code to the extended zone folder
