+++ src/server/scripts/CoreExtended/EasternKingdoms/zone_elwynn_forest.h
